requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.4
django-cron==0.6.0
djangorestframework==3.14.0
filelock==3.8.0
platformdirs==2.5.2
pytz==2022.7
sqlparse==0.4.3
virtualenv==20.16.5
